+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `cli/help`: Fix exit code
 - `cli`: Add missing help options
 - `cli/aux/toggle`: Fix toggle logic if no toggle found
+- `cli/cd`: Fix alias mixed with task IDs
 
 
 ## [0.2.1] - (2026-01-25)